requirements.txt
paho-mqtt==1.6.1
adafruit-circuitpython-dht==4.0.3
gpiozero==2.0.1
picamera2==0.3.16
opencv-python==4.8.1.78
board==1.0
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
psycopg2-binary==2.9.9
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
python-dotenv==1.0.0
pytz==2023.3